tests running again after moving test support files
