+++ package.json
@@ -12,8 +12,14 @@
     "babel-polyfill": "^6.26.0",
     "classnames": "^2.2.5",
     "codemirror": "5.42.2",
+    "diff": "^4.0.1",
     "event-emitter": "^0.3.5",
+    "icss-utils": "^4.0.0",
     "object-hash": "^1.3.1",
+    "postcss": "^7.0.8",
+    "postcss-modules-extract-imports": "^2.0.0",
+    "postcss-modules-scope": "^2.0.1",
+    "postcss-modules-values": "^2.0.0",
     "prop-types": "^15.6.1",
     "react": "16.7.0",
     "react-codemirror2": "^5.1.0",
